refactor: 画面コンポーネントを分割
